Change when the QApplication is initialized.
Instead of creating it inside the _qbzr_run function of all commands, it
is now created as self._application = QApplication() as part of the
base QBzrCommand functionality.

This is done because we need an app before we install QUIFactory otherwise
if the ui_factory is actually *used* we could get a segfault trying to
create a dialog before the QApplication is actually initialized.