Experiment with using a KnownGraph to implement merge_sort.
Very initial results actually aren't great. Time without --lsprof seems
to go from 6.8s to 7.1s to display the MySQL tree. I don't know why, but my
initial guess is that populating the KG is a bit expensive.

Things look better under --lsprof, but probably this is just because it
doesn't actually add the profiling hooks to as much code. lsprof claims
that it is taking 265ms to actually sort the graph, which is probably
accurate. Creating the KnownGraph object seems to take 377ms, which
is certainly longer than sorting. 377ms+265ms is probably longer than
the time for the pure-python merge_sort implementation (without lsprof).

So for it to be a win, we probably need to keep KG around, and do
other work with it, where it can amortize that extra overhead.

I think GaryvdM had done work to use get_known_graph() instead of
doing his own 'get_parent_map' calls. Which should be a net win.